requirements.txt
pillow
requests
json
